% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_DBH.R
\name{fix_DBH}
\alias{fix_DBH}
\alias{mode0}
\alias{fix_height}
\title{Fix DBH}
\usage{
fix_DBH(ddbh, dbh, fun = c("median", "mode"))

mode0(v)

fix_height(height, max_h = 5, sd0 = 0.25)
}
\description{
Fixes negative DBHs by applying a `truncnorm` adjustment
}
