% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_competition.R
\name{light_competition}
\alias{light_competition}
\alias{calc_light_index}
\alias{fix_edge}
\title{Light competition calculation function}
\usage{
calc_light_index(x, y, z, cz = 5, maxx = 200, maxy = 200)

fix_edge()
}
\description{
NA
}
