% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_functions}
\alias{plot_functions}
\alias{HighResTiff}
\alias{lm_eqn}
\alias{Multiplot}
\title{Plot functions for ecohydro2 analysis}
\usage{
HighResTiff(plot_obj, file, width_in, height_in, resolution_dpi, ...)

lm_eqn(df0, xvar, yvar)

Multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\description{
TODO
}
